requirements.txt
requests>=2.31.0
pydantic>=2.6.0
scipy>=1.12.0
numpy>=1.26.0
pandas>=2.2.0
streamlit>=1.32.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-cov>=4.1.0
rich>=13.7.0
nba_api>=1.4.0
